requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Django==4.2.24
whitenoise==6.6.0
psycopg2-binary==2.9.9